use slice in transactions reducer
