+++ src/hooks/useMonitoringEventCallback.ts
@@ -3,7 +3,7 @@ import { initializeApp } from 'firebase/app'
 import { getDatabase, push, ref } from 'firebase/database'
 import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import { useCallback } from 'react'
-import { TransactionInfo, TransactionType } from 'state/transactions/actions'
+import { TransactionInfo, TransactionType } from 'state/transactions/reducer'
 
 type PartialTransactionResponse = Pick<TransactionResponse, 'hash' | 'v' | 'r' | 's'>
