+++ src/pages/MigrateV2/MigrateV2Pair.tsx
@@ -44,7 +44,7 @@ import { useToken } from '../../hooks/Tokens'
 import { usePairContract, useV2MigratorContract } from '../../hooks/useContract'
 import useIsArgentWallet from '../../hooks/useIsArgentWallet'
 import { useTotalSupply } from '../../hooks/useTotalSupply'
-import { TransactionType } from '../../state/transactions/actions'
+import { TransactionType } from '../../state/transactions/reducer'
 import { useTokenBalance } from '../../state/wallet/hooks'
 import { BackArrow, ExternalLink, ThemedText } from '../../theme'
 import { isAddress } from '../../utils'
