+++ src/state/governance/hooks.ts
@@ -32,8 +32,8 @@ import {
 } from '../../constants/proposals'
 import { UNI } from '../../constants/tokens'
 import { useLogs } from '../logs/hooks'
-import { TransactionType } from '../transactions/actions'
 import { useTransactionAdder } from '../transactions/hooks'
+import { TransactionType } from '../transactions/reducer'
 import { VoteOption } from './types'
 
 function useGovernanceV0Contract(): Contract | null {
