+++ src/state/transactions/reducer.ts
@@ -1,16 +1,178 @@
-import { createReducer } from '@reduxjs/toolkit'
+import { createSlice } from '@reduxjs/toolkit'
+import { TradeType } from '@uniswap/sdk-core'
 
 import { updateVersion } from '../global/actions'
-import {
-  addTransaction,
-  checkedTransaction,
-  clearAllTransactions,
-  finalizeTransaction,
-  SerializableTransactionReceipt,
-  TransactionInfo,
-} from './actions'
+import { VoteOption } from '../governance/types'
 
 const now = () => new Date().getTime()
+
+interface SerializableTransactionReceipt {
+  to: string
+  from: string
+  contractAddress: string
+  transactionIndex: number
+  blockHash: string
+  transactionHash: string
+  blockNumber: number
+  status?: number
+}
+
+/**
+ * Be careful adding to this enum, always assign a unique value (typescript will not prevent duplicate values).
+ * These values is persisted in state and if you change the value it will cause errors
+ */
+export enum TransactionType {
+  APPROVAL = 0,
+  SWAP = 1,
+  DEPOSIT_LIQUIDITY_STAKING = 2,
+  WITHDRAW_LIQUIDITY_STAKING = 3,
+  CLAIM = 4,
+  VOTE = 5,
+  DELEGATE = 6,
+  WRAP = 7,
+  CREATE_V3_POOL = 8,
+  ADD_LIQUIDITY_V3_POOL = 9,
+  ADD_LIQUIDITY_V2_POOL = 10,
+  MIGRATE_LIQUIDITY_V3 = 11,
+  COLLECT_FEES = 12,
+  REMOVE_LIQUIDITY_V3 = 13,
+  SUBMIT_PROPOSAL = 14,
+}
+
+export interface BaseTransactionInfo {
+  type: TransactionType
+}
+
+export interface VoteTransactionInfo extends BaseTransactionInfo {
+  type: TransactionType.VOTE
+  governorAddress: string
+  proposalId: number
+  decision: VoteOption
+  reason: string
+}
+
+export interface DelegateTransactionInfo extends BaseTransactionInfo {
+  type: TransactionType.DELEGATE
+  delegatee: string
+}
+
+export interface ApproveTransactionInfo extends BaseTransactionInfo {
+  type: TransactionType.APPROVAL
+  tokenAddress: string
+  spender: string
+}
+
+interface BaseSwapTransactionInfo extends BaseTransactionInfo {
+  type: TransactionType.SWAP
+  tradeType: TradeType
+  inputCurrencyId: string
+  outputCurrencyId: string
+}
+
+export interface ExactInputSwapTransactionInfo extends BaseSwapTransactionInfo {
+  tradeType: TradeType.EXACT_INPUT
+  inputCurrencyAmountRaw: string
+  expectedOutputCurrencyAmountRaw: string
+  minimumOutputCurrencyAmountRaw: string
+}
+export interface ExactOutputSwapTransactionInfo extends BaseSwapTransactionInfo {
+  tradeType: TradeType.EXACT_OUTPUT
+  outputCurrencyAmountRaw: string
+  expectedInputCurrencyAmountRaw: string
+  maximumInputCurrencyAmountRaw: string
+}
+
+export interface DepositLiquidityStakingTransactionInfo {
+  type: TransactionType.DEPOSIT_LIQUIDITY_STAKING
+  token0Address: string
+  token1Address: string
+}
+
+export interface WithdrawLiquidityStakingTransactionInfo {
+  type: TransactionType.WITHDRAW_LIQUIDITY_STAKING
+  token0Address: string
+  token1Address: string
+}
+
+export interface WrapTransactionInfo {
+  type: TransactionType.WRAP
+  unwrapped: boolean
+  currencyAmountRaw: string
+  chainId?: number
+}
+
+export interface ClaimTransactionInfo {
+  type: TransactionType.CLAIM
+  recipient: string
+  uniAmountRaw?: string
+}
+
+export interface CreateV3PoolTransactionInfo {
+  type: TransactionType.CREATE_V3_POOL
+  baseCurrencyId: string
+  quoteCurrencyId: string
+}
+
+export interface AddLiquidityV3PoolTransactionInfo {
+  type: TransactionType.ADD_LIQUIDITY_V3_POOL
+  createPool: boolean
+  baseCurrencyId: string
+  quoteCurrencyId: string
+  feeAmount: number
+  expectedAmountBaseRaw: string
+  expectedAmountQuoteRaw: string
+}
+
+export interface AddLiquidityV2PoolTransactionInfo {
+  type: TransactionType.ADD_LIQUIDITY_V2_POOL
+  baseCurrencyId: string
+  quoteCurrencyId: string
+  expectedAmountBaseRaw: string
+  expectedAmountQuoteRaw: string
+}
+
+export interface MigrateV2LiquidityToV3TransactionInfo {
+  type: TransactionType.MIGRATE_LIQUIDITY_V3
+  baseCurrencyId: string
+  quoteCurrencyId: string
+  isFork: boolean
+}
+
+export interface CollectFeesTransactionInfo {
+  type: TransactionType.COLLECT_FEES
+  currencyId0: string
+  currencyId1: string
+}
+
+export interface RemoveLiquidityV3TransactionInfo {
+  type: TransactionType.REMOVE_LIQUIDITY_V3
+  baseCurrencyId: string
+  quoteCurrencyId: string
+  expectedAmountBaseRaw: string
+  expectedAmountQuoteRaw: string
+}
+
+export interface SubmitProposalTransactionInfo {
+  type: TransactionType.SUBMIT_PROPOSAL
+}
+
+export type TransactionInfo =
+  | ApproveTransactionInfo
+  | ExactOutputSwapTransactionInfo
+  | ExactInputSwapTransactionInfo
+  | ClaimTransactionInfo
+  | VoteTransactionInfo
+  | DelegateTransactionInfo
+  | DepositLiquidityStakingTransactionInfo
+  | WithdrawLiquidityStakingTransactionInfo
+  | WrapTransactionInfo
+  | CreateV3PoolTransactionInfo
+  | AddLiquidityV3PoolTransactionInfo
+  | AddLiquidityV2PoolTransactionInfo
+  | MigrateV2LiquidityToV3TransactionInfo
+  | CollectFeesTransactionInfo
+  | RemoveLiquidityV3TransactionInfo
+  | SubmitProposalTransactionInfo
 
 export interface TransactionDetails {
   hash: string
@@ -30,9 +192,44 @@ export interface TransactionState {
 
 export const initialState: TransactionState = {}
 
-export default createReducer(initialState, (builder) =>
-  builder
-    .addCase(updateVersion, (transactions) => {
+const transactionSlice = createSlice({
+  name: 'transactions',
+  initialState,
+  reducers: {
+    addTransaction(transactions, { payload: { chainId, from, hash, info } }) {
+      if (transactions[chainId]?.[hash]) {
+        throw Error('Attempted to add existing transaction.')
+      }
+      const txs = transactions[chainId] ?? {}
+      txs[hash] = { hash, info, from, addedTime: now() }
+      transactions[chainId] = txs
+    },
+    clearAllTransactions(transactions, { payload: { chainId } }) {
+      if (!transactions[chainId]) return
+      transactions[chainId] = {}
+    },
+    checkedTransaction(transactions, { payload: { chainId, hash, blockNumber } }) {
+      const tx = transactions[chainId]?.[hash]
+      if (!tx) {
+        return
+      }
+      if (!tx.lastCheckedBlockNumber) {
+        tx.lastCheckedBlockNumber = blockNumber
+      } else {
+        tx.lastCheckedBlockNumber = Math.max(blockNumber, tx.lastCheckedBlockNumber)
+      }
+    },
+    finalizeTransaction(transactions, { payload: { hash, chainId, receipt } }) {
+      const tx = transactions[chainId]?.[hash]
+      if (!tx) {
+        return
+      }
+      tx.receipt = receipt
+      tx.confirmedTime = now()
+    },
+  },
+  extraReducers: (builder) => {
+    builder.addCase(updateVersion, (transactions) => {
       // in case there are any transactions in the store with the old format, remove them
       Object.keys(transactions).forEach((chainId) => {
         const chainTransactions = transactions[chainId as unknown as number]
@@ -44,35 +241,9 @@ export default createReducer(initialState, (builder) =>
         })
       })
     })
-    .addCase(addTransaction, (transactions, { payload: { chainId, from, hash, info } }) => {
-      if (transactions[chainId]?.[hash]) {
-        throw Error('Attempted to add existing transaction.')
-      }
-      const txs = transactions[chainId] ?? {}
-      txs[hash] = { hash, info, from, addedTime: now() }
-      transactions[chainId] = txs
-    })
-    .addCase(clearAllTransactions, (transactions, { payload: { chainId } }) => {
-      if (!transactions[chainId]) return
-      transactions[chainId] = {}
-    })
-    .addCase(checkedTransaction, (transactions, { payload: { chainId, hash, blockNumber } }) => {
-      const tx = transactions[chainId]?.[hash]
-      if (!tx) {
-        return
-      }
-      if (!tx.lastCheckedBlockNumber) {
-        tx.lastCheckedBlockNumber = blockNumber
-      } else {
-        tx.lastCheckedBlockNumber = Math.max(blockNumber, tx.lastCheckedBlockNumber)
-      }
-    })
-    .addCase(finalizeTransaction, (transactions, { payload: { hash, chainId, receipt } }) => {
-      const tx = transactions[chainId]?.[hash]
-      if (!tx) {
-        return
-      }
-      tx.receipt = receipt
-      tx.confirmedTime = now()
-    })
-)
+  },
+})
+
+export const { addTransaction, clearAllTransactions, checkedTransaction, finalizeTransaction } =
+  transactionSlice.actions
+export default transactionSlice.reducer
